Re-resolve the DNS name of the Redis server before reconnecting
This solves the case where the DNS record of the Redis server is
updated and the connection breaks. The previous code would try the same
IP address over and over, we now re-create the InetSocketAddress object
to force re-resolving.

Note that the JVM will do some DNS caching on its own (and the
operating system potentially too), this commit just fixes the issue at
library level.